Move SdkConfiguration into redux store
