+++ src/components/Router/MainRouter.tsx
@@ -10,7 +10,7 @@ import type { MobileConfig } from '~types/commons'
 import type { StepConfig } from '~types/steps'
 import type { FlowChangeCallback, InternalRouterProps } from '~types/routers'
 import Spinner from '../Spinner'
-import { SdkConfigurationServiceProvider } from '~contexts/useSdkConfigurationService'
+import { SdkConfigurationServiceProvider } from '~core/SdkConfiguration/useSdkConfigurationService'
 import { createOptionsStepsHook } from './createOptionsStepsHook'
 import { createWorkflowStepsHook } from './createWorkflowStepsHook'
 import { UserConsentProvider } from '~contexts/useUserConsent'
