+++ src/components/utils/onfidoApi.ts
@@ -18,7 +18,6 @@ import type {
   CreateV4DocumentResponse,
   SuccessCallback,
   ErrorCallback,
-  SdkConfiguration,
   ApplicantConsent,
   ApplicantConsentStatus,
   PoASupportedCountry,
@@ -33,6 +32,7 @@ import type {
   LegacyTrackedEventNames,
 } from '~types/tracker'
 import type { DocumentTypes, PoaTypes } from '~types/steps'
+import { SdkConfiguration } from '~core/SdkConfiguration/types'
 
 export type UploadPayload = {
   filename?: string
@@ -665,39 +665,3 @@ export const sendAnalytics = (
     }
   )
 }
-
-export const getSdkConfiguration = (
-  url: string,
-  token: string
-): Promise<SdkConfiguration> =>
-  new Promise((resolve, reject) => {
-    try {
-      const browserInfo = detectSystem('browser')
-      const osInfo = detectSystem('os')
-
-      const requestParams: HttpRequestParams = {
-        endpoint: `${url}/v3.3/sdk/configurations`,
-        token: `Bearer ${token}`,
-        contentType: 'application/json',
-        method: 'POST',
-        payload: JSON.stringify({
-          sdk_source: process.env.SDK_SOURCE,
-          sdk_version: process.env.SDK_VERSION,
-          sdk_metadata: {
-            system: {
-              browser: browserInfo.name,
-              browser_version: browserInfo.version,
-              os: osInfo.name,
-              os_version: osInfo.version,
-            },
-          },
-        }),
-      }
-
-      performHttpRequest(requestParams, resolve, (request) =>
-        formatError(request, reject)
-      )
-    } catch (error) {
-      reject(error)
-    }
-  })
