+++ src/core/SdkConfiguration/config.ts
@@ -1,0 +1,17 @@
+import { SdkConfiguration } from './types'
+
+export const defaultConfiguration: SdkConfiguration = {
+  experimental_features: {
+    enable_multi_frame_capture: false,
+    motion_experiment: {
+      enabled: false,
+    },
+  },
+  sdk_features: {
+    enable_require_applicant_consents: true,
+    disable_cross_device_sms: false,
+  },
+  document_capture: {
+    max_total_retries: 1,
+  },
+}
