+++ src/types/redux/constants.ts
@@ -36,5 +36,6 @@ export const SET_IS_CROSS_DEVICE_CLIENT = 'SET_IS_CROSS_DEVICE_CLIENT'
 export const CAPTURE_CREATE = 'CAPTURE_CREATE'
 export const CAPTURE_DELETE = 'CAPTURE_DELETE'
 export const SET_CAPTURE_METADATA = 'SET_CAPTURE_METADATA'
+export const SET_SDK_CONFIGURATION = 'SET_SDK_CONFIGURATION'
 
 export const RESET_STORE = 'RESET_STORE'
